Rename SanEncoder to PgnNodeEncoder
And other refactor & comments
